remove unnecessary cast

+++ groups/bae/baenet/baenet_httpgeneratorutil.cpp
@@ -497,7 +497,7 @@ int baenet_HttpGeneratorUtil::generateHeader(
        << '.' 
        << statusLine.minorVersion()
        << ' ' 
-       << (int)statusLine.statusCode()
+       << statusLine.statusCode()
        << ' ' 
        << statusLine.reasonPhrase() 
        << "\r\n";
